refactor: 머지 충돌 확인 및 이름 변경
