+++ src/controller/applicationController.ts
@@ -8,7 +8,7 @@ export const createApplicationData = async (
   next: NextFunction,
 ) => {
   try {
-    let applyData = req.body;      //POST 방식으로 정보를 받아 온다.
+    let applyData = req.body; //POST 방식으로 정보를 받아 온다.
 
     await applicationService.createApplicationData(applyData);
     res.status(200).send();
@@ -23,8 +23,8 @@ export const getApplicationData = async (
   next: NextFunction,
 ) => {
   try {
-    const userId = new Types.ObjectId("64c28df8f3ad2b9ac18163e1");  //임의로 설정하였고, 원래대로라면 request에서 받아야 함.
-  
+    const userId = new Types.ObjectId('64c76faea01a83ef15a9f141'); //임의로 설정하였고, 원래대로라면 request에서 받아야 함.
+
     const userData = await applicationService.getApplicationData(userId); //user Data를 찾아서 보낸다.
     res.send(userData);
   } catch (err) {
@@ -38,8 +38,8 @@ export const deleteApplicationData = async (
   next: NextFunction,
 ) => {
   try {
-    const userId = new Types.ObjectId("64c28df8f3ad2b9ac18163e1");  //임의로 설정하였고, 원래대로라면 request에서 받아야 함.
-  
+    const userId = new Types.ObjectId('64c28df8f3ad2b9ac18163e1'); //임의로 설정하였고, 원래대로라면 request에서 받아야 함.
+
     const userData = await applicationService.deleteApplicationData(userId); //user Data를 찾아서 보낸다.
     res.send(userData);
   } catch (err) {
@@ -53,7 +53,7 @@ export const updateApplicationData = async (
   next: NextFunction,
 ) => {
   try {
-    let applyData = req.body;      //POST 방식으로 정보를 받아 온다.
+    let applyData = req.body; //POST 방식으로 정보를 받아 온다.
 
     await applicationService.updateApplicationData(applyData);
     res.status(200).send();
